+++ src/bin/conserve.rs
@@ -145,6 +145,9 @@ enum Command {
         /// Show only version names.
         #[structopt(long, short = "q")]
         short: bool,
+        /// Sort bands to show most recent first.
+        #[structopt(long, short = "n")]
+        newest: bool,
         /// Show size of stored trees.
         #[structopt(long, short = "z", conflicts_with = "short")]
         sizes: bool,
@@ -357,15 +360,22 @@ impl Command {
             Command::Versions {
                 archive,
                 short,
+                newest,
                 sizes,
                 utc,
             } => {
                 ui::enable_progress(false);
                 let archive = Archive::open_path(archive)?;
                 if *short {
-                    output::show_brief_version_list(&archive, &mut stdout)?;
+                    output::show_brief_version_list(&archive, *newest, &mut stdout)?;
                 } else {
-                    output::show_verbose_version_list(&archive, *sizes, *utc, &mut stdout)?;
+                    output::show_verbose_version_list(
+                        &archive,
+                        *sizes,
+                        *utc,
+                        *newest,
+                        &mut stdout,
+                    )?;
                 }
             }
         }